fix: optimize array merging for associative translations, cnsure proper type casting for AI config values

+++ config/auto-translations.php
@@ -54,9 +54,9 @@ return [
         'chatgpt' => [
             'api_key' => env('CHATGPT_API_KEY'),
             'model' => env('CHATGPT_MODEL', 'gpt-3.5-turbo'),
-            'temperature' => env('CHATGPT_TEMPERATURE', 0.7),
-            'max_tokens' => env('CHATGPT_MAX_TOKENS', 4096),
-            'http_timeout' => env('CHATGPT_HTTP_TIMEOUT', 300),
+            'temperature' => (float) env('CHATGPT_TEMPERATURE', 0.7),
+            'max_tokens' => (int) env('CHATGPT_MAX_TOKENS', 4096),
+            'http_timeout' => (int) env('CHATGPT_HTTP_TIMEOUT', 300),
         ],
 
         'google' => [
